+++ noggin/tests/unit/controller/test_forgot_password.py
@@ -295,9 +295,13 @@ def test_change_post_generic_error(
 
 @pytest.mark.vcr()
 def test_change_post_with_otp(
-    client, dummy_user, dummy_user_with_otp, token_for_dummy_user, patched_lock_active
+    client,
+    dummy_user,
+    logged_in_dummy_user_with_otp,
+    token_for_dummy_user,
+    patched_lock_active,
 ):
-    otp = get_otp(otp_secret_from_uri(dummy_user_with_otp.uri))
+    otp = get_otp(otp_secret_from_uri(logged_in_dummy_user_with_otp.uri))
     with fml_testing.mock_sends(
         UserUpdateV1(
             {"msg": {"agent": "dummy", "user": "dummy", "fields": ["password"]}}
@@ -324,7 +328,7 @@ def test_change_post_with_otp(
 def test_change_post_password_with_otp_not_given(
     client,
     dummy_user,
-    dummy_user_with_otp,
+    logged_in_dummy_user_with_otp,
     token_for_dummy_user,
     patched_lock_active,
     mocker,
@@ -346,7 +350,7 @@ def test_change_post_password_with_otp_not_given(
 def test_change_post_password_with_otp_wrong_value(
     client,
     dummy_user,
-    dummy_user_with_otp,
+    logged_in_dummy_user_with_otp,
     token_for_dummy_user,
     patched_lock_active,
     mocker,