+++ noggin/tests/unit/controller/test_user_otp.py
@@ -12,17 +12,19 @@ from noggin.tests.unit.utilities import (
     assert_form_field_error,
     assert_form_generic_error,
     assert_redirects_with_flash,
+    get_otp,
+    otp_secret_from_uri,
 )
 
 
 @pytest.fixture
-def dummy_user_with_2_otp(client, logged_in_dummy_user, dummy_user_with_otp):
+def dummy_user_with_2_otp(client, logged_in_dummy_user, logged_in_dummy_user_with_otp):
     ipa = logged_in_dummy_user
     result = ipa.otptoken_add(
         o_ipatokenowner="dummy", o_description="dummy's other token",
     )['result']
     token = OTPToken(result)
-    yield dummy_user_with_otp, token
+    yield logged_in_dummy_user_with_otp, token
     try:
         ipa_admin.otptoken_del(token.uniqueid)
     except python_freeipa.exceptions.NotFound:
@@ -147,7 +149,45 @@ def test_user_settings_otp_confirm(
 
 @pytest.mark.vcr()
 def test_user_settings_otp_add_second(
-    client, dummy_user_with_otp, logged_in_dummy_user, cleanup_dummy_tokens, totp_token
+    client, logged_in_dummy_user_with_otp, cleanup_dummy_tokens
+):
+    """Test posting to the create OTP endpoint"""
+    otp = get_otp(otp_secret_from_uri(logged_in_dummy_user_with_otp.uri))
+    result = client.post(
+        "/user/dummy/settings/otp/",
+        data={
+            "add-description": "pants token 2",
+            "add-password": "dummy_password",
+            "add-otp": otp,
+            "add-submit": "1",
+        },
+    )
+    page = BeautifulSoup(result.data, "html.parser")
+    tokenlist = page.select_one("div.list-group")
+    assert tokenlist is not None
+    tokens = tokenlist.select(".list-group-item div[data-role='token-description']")
+    assert len(tokens) == 1
+
+    modal = page.select_one("#otp-modal")
+    assert modal is not None
+
+    confirm_form = modal.select_one("form")
+    assert confirm_form is not None
+    assert (
+        confirm_form.select_one("input[name='confirm-description']")["value"]
+        == "pants token 2"
+    )
+    otp_uri = page.select_one("input#otp-uri")
+    parsed_otp_uri_query = parse_qs(urlparse(otp_uri["value"]).query)
+    assert (
+        confirm_form.select_one("input[name='confirm-secret']")["value"]
+        == parsed_otp_uri_query["secret"][0]
+    )
+
+
+@pytest.mark.vcr()
+def test_user_settings_otp_add_second_confirm(
+    client, logged_in_dummy_user_with_otp, cleanup_dummy_tokens, totp_token,
 ):
     """Test posting to the create OTP endpoint"""
     result = client.post(
@@ -386,13 +426,11 @@ def test_user_settings_otp_disable(client, logged_in_dummy_user, dummy_user_with
 
 
 @pytest.mark.vcr()
-def test_user_settings_otp_disable_lasttoken(
-    client, logged_in_dummy_user, dummy_user_with_otp
-):
+def test_user_settings_otp_disable_lasttoken(client, logged_in_dummy_user_with_otp):
     """Test trying to disable the last token"""
     result = client.post(
         "/user/dummy/settings/otp/disable/",
-        data={"token": dummy_user_with_otp.uniqueid},
+        data={"token": logged_in_dummy_user_with_otp.uniqueid},
     )
     assert_redirects_with_flash(
         result,
@@ -524,7 +562,7 @@ def test_user_settings_otp_delete(client, logged_in_dummy_user, dummy_user_with_
 
 @pytest.mark.vcr()
 def test_user_settings_otp_delete_lasttoken(
-    client, logged_in_dummy_user, dummy_user_with_otp
+    client, logged_in_dummy_user, logged_in_dummy_user_with_otp
 ):
     """Test trying to delete the last token"""
     result = client.get("/user/dummy/settings/otp/")